logout and endpoint

+++ src/main/java/kskowronski/data/service/egeria/CashRegisterRepo.java
@@ -3,5 +3,7 @@ package kskowronski.data.service.egeria;
 import kskowronski.data.entity.egeria.CashRegister;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CashRegisterRepo extends JpaRepository<CashRegister, Integer> {
+import java.math.BigDecimal;
+
+public interface CashRegisterRepo extends JpaRepository<CashRegister, BigDecimal> {
 }
